requirements.txt
torch>=2.0.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
owslib>=0.29.0
requests>=2.31.0
tk>=0.1.0
tkcalendar>=1.6.1